refactor: Rename Profile component to FolderInfo
