+++ src/components/FolderInfo/FolderInfo.js
